+++ src/components/Profile.tsx
@@ -8,7 +8,10 @@ import {
   formatBalance,
   formatPrimaryName,
   formatWalletAddress,
+  getBlockExplorerUrlForAddress,
+  isEthAddress,
 } from '@src/utils';
+import { SendHorizonal, WalletMinimal } from 'lucide-react';
 import { forwardRef, ReactElement, useState } from 'react';
 import Button, { ButtonType } from './Button';
 import CopyButton from './CopyButton';
@@ -21,7 +24,7 @@ import {
   LogoutIcon,
 } from './icons';
 import ConnectModal from './modals/ConnectModal';
-import { WalletMinimal } from 'lucide-react';
+import TransferArioModal from './modals/TransferArioModal';
 
 // eslint-disable-next-line react/display-name
 const CustomPopoverButton = forwardRef<
@@ -58,82 +61,108 @@ const Profile = () => {
   const { data: balances } = useBalances(walletAddress);
   const ticker = useGlobalState((state) => state.ticker);
 
-  const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
+  const [showConnectModal, setShowConnectModal] = useState<boolean>(false);
+  const [showTransferModal, setShowTransferModal] = useState<boolean>(false);
   const { data: primaryName } = usePrimaryName(walletAddress?.toString());
   const { data: logo } = useLogo({ primaryName: primaryName?.name });
 
   return walletAddress ? (
     <Popover className="relative">
-      <PopoverButton as={CustomPopoverButton} logo={logo}>
-        {primaryName
-          ? formatPrimaryName(primaryName.name)
-          : formatWalletAddress(walletAddress.toString())}
-      </PopoverButton>
+      {({ close }) => (
+        <>
+          <PopoverButton as={CustomPopoverButton} logo={logo}>
+            {primaryName
+              ? formatPrimaryName(primaryName.name)
+              : formatWalletAddress(walletAddress.toString())}
+          </PopoverButton>
 
-      <PopoverPanel className="absolute right-0 z-50 mt-2.5 w-fit rounded-xl border border-grey-800 bg-grey-1000 text-sm shadow-xl">
-        <div className="flex gap-2 px-4 py-5 ">
-          <WalletMinimal className="size-4" />
+          <PopoverPanel className="absolute right-0 z-50 mt-2.5 w-fit rounded-xl border border-grey-800 bg-grey-1000 text-sm shadow-xl min-w-52">
+            <div className="flex gap-2 px-4 py-5 ">
+              <WalletMinimal className="size-4" />
 
-          <div className="flex gap-2 align-middle text-mid">
-            <a
-              href={`https://viewblock.io/arweave/address/${walletAddress.toString()}`}
-              target="_blank"
-              rel="noreferrer"
-              onClick={(e) => {
-                e.stopPropagation();
-              }}
-            >
-              <Tooltip
-                message={
-                  <div className="text-high">{walletAddress.toString()}</div>
-                }
-                useMaxWidth={false}
+              <div className="flex gap-2 align-middle text-mid">
+                <a
+                  href={getBlockExplorerUrlForAddress(walletAddress.toString())}
+                  target="_blank"
+                  rel="noreferrer"
+                  onClick={(e) => {
+                    e.stopPropagation();
+                  }}
+                >
+                  <Tooltip
+                    message={
+                      <div className="text-high">
+                        {walletAddress.toString()}
+                      </div>
+                    }
+                    useMaxWidth={false}
+                  >
+                    {formatWalletAddress(walletAddress.toString())}
+                  </Tooltip>
+                </a>
+              </div>
+              <CopyButton textToCopy={walletAddress.toString()} />
+            </div>
+            <div className="mx-4  rounded-md border border-grey-800 py-3">
+              <div className="relative border-b border-grey-800">
+                <div className="px-4 text-xs text-low">{ticker} Balance</div>
+                <div className="px-4 pb-3 pt-1 text-high">
+                  {balances ? formatBalance(balances.ario) : <Placeholder />}
+                </div>
+                {balances?.ario && (
+                  <button
+                    className="absolute right-4 top-1/2 -translate-y-5 rounded border border-grey-800 p-2"
+                    title={`Send ${ticker}`}
+                    onClick={() => {
+                      setShowTransferModal(true);
+                      close();
+                    }}
+                  >
+                    <SendHorizonal className="size-3" />
+                  </button>
+                )}
+              </div>
+              <div className="px-4 pt-3 text-xs text-low">AR Balance</div>
+              <div className="px-4 pt-1 text-high">
+                {balances ? formatBalance(balances.ar) : <Placeholder />}
+              </div>
+            </div>
+            {!isEthAddress(walletAddress.toString()) && (
+              <div className="flex flex-col gap-3 text-nowrap px-6 pt-3 text-mid">
+                <button
+                  className="flex items-center"
+                  title="Transaction History"
+                  onClick={async () => {
+                    window.open(
+                      `https://ao.link/#/entity/${walletAddress.toString()}`,
+                      '_blank',
+                    );
+                  }}
+                >
+                  <ClockRewindIcon className="mr-2 h-4 w-[.9375rem]" />{' '}
+                  Transaction History
+                  <LinkArrowIcon className="ml-1 size-3" />
+                </button>
+              </div>
+            )}
+            <div className="mt-3 flex flex-col gap-3 bg-btn-secondary-default px-6 py-3 text-mid">
+              <button
+                className="flex items-center gap-2"
+                title="Logout"
+                onClick={async () => {
+                  await wallet?.disconnect();
+                  updateWallet(undefined, undefined);
+                }}
               >
-                {formatWalletAddress(walletAddress.toString())}
-              </Tooltip>
-            </a>
-          </div>
-          <CopyButton textToCopy={walletAddress.toString()} />
-        </div>
-        <div className="mx-4  rounded-md border border-grey-800 py-3">
-          <div className="px-4 text-xs text-low">{ticker} Balance</div>
-          <div className="border-b border-grey-800 px-4 pb-3 pt-1 text-high">
-            {balances ? formatBalance(balances.io) : <Placeholder />}
-          </div>
-          <div className="px-4 pt-3 text-xs text-low">AR Balance</div>
-          <div className="px-4 pt-1 text-high">
-            {balances ? formatBalance(balances.ar) : <Placeholder />}
-          </div>
-        </div>
-        <div className="flex flex-col gap-3 text-nowrap px-6 py-3 text-mid">
-          <button
-            className="flex items-center"
-            title="Transaction History"
-            onClick={async () => {
-              window.open(
-                `https://ao.link/#/entity/${walletAddress.toString()}`,
-                '_blank',
-              );
-            }}
-          >
-            <ClockRewindIcon className="mr-2 h-4 w-[.9375rem]" /> Transaction
-            History
-            <LinkArrowIcon className="ml-1 size-3" />
-          </button>
-        </div>
-        <div className="flex flex-col gap-3 bg-btn-secondary-default px-6 py-3 text-mid">
-          <button
-            className="flex items-center gap-2"
-            title="Logout"
-            onClick={async () => {
-              await wallet?.disconnect();
-              updateWallet(undefined, undefined);
-            }}
-          >
-            <LogoutIcon className="size-4" /> Logout
-          </button>
-        </div>
-      </PopoverPanel>
+                <LogoutIcon className="size-4" /> Logout
+              </button>
+            </div>
+          </PopoverPanel>
+          {showTransferModal && (
+            <TransferArioModal onClose={() => setShowTransferModal(false)} />
+          )}
+        </>
+      )}
     </Popover>
   ) : walletStateInitialized ? (
     <div>
@@ -142,9 +171,11 @@ const Profile = () => {
         icon={<ConnectIcon className="size-4" />}
         title="Connect"
         text="Connect"
-        onClick={() => setIsModalOpen(true)}
+        onClick={() => setShowConnectModal(true)}
       />
-      {isModalOpen && <ConnectModal onClose={() => setIsModalOpen(false)} />}
+      {showConnectModal && (
+        <ConnectModal onClose={() => setShowConnectModal(false)} />
+      )}
     </div>
   ) : (
     <div></div>