+++ src/utils/errors.ts
@@ -13,10 +13,17 @@ export class ValidationError extends NotificationOnlyError {
   }
 }
 
-export class ArconnectError extends NotificationOnlyError {
+export class WanderError extends NotificationOnlyError {
   constructor(message: string) {
     super(message);
-    this.name = 'ArConnect';
+    this.name = 'Wander';
+  }
+}
+
+export class MetamaskError extends NotificationOnlyError {
+  constructor(message: string) {
+    super(message);
+    this.name = 'Metamask';
   }
 }